backend-python/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pandas
openpyxl
python-multipart
requests
httpx
pydantic-settings
aiofiles
sqlglot>=23.0.0
psycopg[binary]>=3.1.18
cryptography>=41.0.0
# Optional LLM providers (uncomment if keys configured)
groq>=0.9.0
google-generativeai>=0.7.0
openai>=1.40.0
